requirements.txt
azureml-sdk
streamlit
pandas
numpy
scikit-learn
plotly
pybase64
pymongo
pytz
bcrypt
pymongo[srv]
pydot
graphviz
connect
statsmodels
bs4
nltk
matplotlib
xlrd
openpyxl


